+++ gym_chrono/envs/.gitattributes
@@ -1,0 +1,1 @@
+data/** filter=lfs diff=lfs merge=lfs -text
