+++ gym_chrono/test/art_wpts_test.py
@@ -1,6 +1,6 @@
 import gymnasium as gym
 from stable_baselines3 import A2C, SAC, PPO, TD3
-from gym_chrono.envs.driving.art_wpts import art_wpts
+from gym_chrono.envs.wheeled.art_wpts import art_wpts
 
 env = art_wpts()
